requirements.txt
PyJWT==1.4.2
Flask==0.10.1
requests==2.7.0
Flask-Cors==3.0.2
Flask-Environ